requirements.txt
# Core LangChain
langchain
langchain-core
langchain-community

# Text splitting
langchain-text-splitters

# Vector Store
langchain-chroma
chromadb

# FREE Embeddings (no API needed)
sentence-transformers
huggingface-hub
# langchain-huggingface
langchain-ollama
torch  # Optional, but helps with embeddings

# Utilities
python-dotenv
tqdm

# Optional: Keep Google AI for future use
# langchain-google-genai
# google-generativeai
